[level 1] Title: 크기가 작은 부분문자열, Time: 14.92 ms, Memory: 64.4 MB -BaekjoonHub

+++ 프로그래머스/1/147355. 크기가 작은 부분문자열/크기가 작은 부분문자열.kt
@@ -1,3 +1,3 @@
 class Solution {
-    fun solution(t: String, p: String): Int = t.windowed(p.length).count { it.toLong() <= p.toLong() }
+    fun solution(t: String, p: String): Int = t.windowed(p.length).count { it <= p }
 }
